+++ CHANGELOG.md
@@ -3,10 +3,12 @@ TimesheetPlus - Changelog
 
 V1.1
 ----
-* Improved task command: Updating task number now accepts multiple comma-separated entry IDs 
+* Bugfix: Viewing empty timesheet w/ filter was stuck in loop
+* Improved CLI timesheet view: Exclamation mark HTML entity was not decoded 
+* Improved comment command: Updating/appending to existing comments now accepts ID argument w/o "i=" prefix
 * Improved comment command: Updating/appending to existing comments now accepts multiple comma-separated entry IDs 
 * Improved resume command: When user entered task-number instead entry ID, the last entry of that task is suggested to be resumed
-* Bugfix: Viewing empty timesheet w/ filter was stuck in loop
+* Improved task command: Updating task number now accepts multiple comma-separated entry IDs 
 
 
 v1.0