+++ CHANGELOG.md
@@ -5,6 +5,7 @@ V1.1
 ----
 * Bugfix: Viewing empty timesheet w/ filter was stuck in loop
 * Improved CLI timesheet view: Exclamation mark HTML entity was not decoded 
+* Improved comment command: Appending to existing comment allows text in double quotes to start w/ a space character
 * Improved comment command: Updating/appending to existing comments now accepts ID argument w/o "i=" prefix
 * Improved comment command: Updating/appending to existing comments now accepts multiple comma-separated entry IDs 
 * Improved resume command: When user entered task-number instead entry ID, the last entry of that task is suggested to be resumed